Code-for-Experiment/Targeted-Training/dit_training_on_hifigan_vae/requirements.txt
torch
torchaudio
numpy
scikit-learn
tqdm
accelerate
einops
omegaconf
ffmpeg
tensorboard
gradio
google-auth
transformers==4.36.0
sentencepiece
encodec
demucs
pyloudnorm
laion-clap
tiktoken
sacremoses
librosa
opencc-python-reimplemented
# deepspeed
